Instance Files Reader

Contains methods for reading instances stored in text files. Types of instances include: graphs, set cover problem and directed and undirected Steiner tree problem.

VKSetCovers.create_graph — Function

Create an object of type graphs.Graph given the number of vertices and the list of edges.

Parameters

- num_verts::Int32 - number of vertices in the graph.
- edges::Vector{Tuple{Int32, Int32}} - edges of the graph.

Returns

- ::Graphs.Graph - a Graphs.Graph object representing the graph.

VKSetCovers.read_graph_from_mtx_file — Function

Reads a graph stored in .mxt format in the given file.

Parameters

- input_file::String - string with the file path.

Returns

- num_verts::Int32 - number of vertices in the graph.
- num_edges::Int32 - number of edges in the graph.
- edges::Vector{Tuple{Int32, Int32}} - edges of the graph.

VKSetCovers.get_graph_from_mtx_file — Function

Returns an object of type Graphs.Graph representing the graph stored in the given .mtx format file.

Parameters

- input_file::String - string with the file path.

Returns

- ::Graphs.Graph - a Graphs.Graph object representing the graph.

VKSetCovers.read_set_cover_instance_from_file — Function

Reads a set covering problem data from the given file.

Parameters

- input_file::String - string with the file path.

Returns

- num_ineqs::Int32 - number of inequalities in SCP instance.
- num_vars::Int32 - number of variables in SCP instance.
- var_spanned_ineqs::Vector{Vector{Int32}} - the set of inequalities covered by each variable.
- var_costs::Vector{Int32} - the objective value cost of each variable.

VKSetCovers.get_set_cover_instance_from_file — Function

Returns an object of type SetCoverInstance containing the data of the given set cover instance.

Parameters

- input_file::String - string with the file path.

Returns

- ::SetCoverInstance - object with SCP instance data.

VKSetCovers.read_undirected_steiner_tree_instance_from_file — Function

Reads the data of the undirected Steiner tree problem instance in the given file.

The data corresponds to: graph vertices and edges, edge weights and terminal vertices.

Parameters

- input_file::String - string with the file path.

Returns

- num_verts::Int32 - number of vertices in the graph.
- num_edges::Int32 - number of edges in the graph.
- edges::Vector{Tuple{Int32, Int32}} - the edges of the graph.
- edge_weights::Dict{Tuple{Int32, Int32}, Int32} - the weight for each edge.
- terminal_verts::Vector{Int32} - the set of terminal vertices.

VKSetCovers.read_directed_steiner_tree_instance_from_file — Function

Reads the data of the directed Steiner tree problem instance in the given file.

The data corresponds to: graph vertices and arcs, arc weights and terminal vertices.

Parameters

- input_file::String - string with the file path.

Returns

- num_verts::Int32 - number of vertices in the digraph.
- num_arcs::Int32 - number of arcs in the digraph.
- edges::Vector{Tuple{Int32, Int32}} - the edges of the digraph.
- edge_weights::Dict{Tuple{Int32, Int32}, Int32} - the weight for each arc.
- terminal_verts::Vector{Int32} - the set of terminal vertices.
- root::Int32 - the root vertex.